test for command implementations

+++ EpamFinalProject/src/main/java/com/example/epamfinalproject/Controllers/Commands/Administrator/DeleteCruiseCommand.java
@@ -1,6 +1,7 @@
 package com.example.epamfinalproject.Controllers.Commands.Administrator;
 
 import com.example.epamfinalproject.Controllers.Commands.Command;
+import com.example.epamfinalproject.Controllers.MessageKeys;
 import com.example.epamfinalproject.Controllers.Path;
 import com.example.epamfinalproject.Services.CruiseService;
 import com.example.epamfinalproject.Utility.Constants;
@@ -11,6 +12,11 @@ import javax.servlet.http.HttpServletRequest;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 
+/**
+ * The Command to delete a Cruise record from the database.
+ * The command changes the value of the "deleted" field to true,
+ * after that the record is no longer displayed in the interface
+ */
 public class DeleteCruiseCommand implements Command {
   private static final Logger log = LogManager.getLogger(DeleteCruiseCommand.class);
 
@@ -31,6 +37,11 @@ public class DeleteCruiseCommand implements Command {
       return Constants.REDIRECT + Path.ADMINISTRATOR_PAGE;
     }
     for (String param : params) {
+      if(Integer.parseInt(param) <= 0){
+        log.warn("Invalid Entity ID");
+        request.getSession().setAttribute(Constants.MESSAGE, MessageKeys.PARAM_INVALID);
+        return Constants.REDIRECT + Path.ADMINISTRATOR_PAGE;
+      }
       cruiseService.deleteCruiseByID(Integer.parseInt(param));
       log.debug("Record with id " + Integer.parseInt(param) + "was deleted");
     }
